typer for blanket og brev skal adskilles i egne mapper
