+++ src/server/blankett/components/AlderPåBarnGrunnlag.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import type { IBarnMedSamvær } from '../../../typer/dokumentApi';
+import type { IBarnMedSamvær } from '../../../typer/dokumentApiBlankett';
 import { formaterNullableIsoDato } from '../../utils/util';
 
 interface Props {
