+++ src/server/blankett/components/MorEllerFarGrunnlag.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import type { IBarnMedSamvær } from '../../../typer/dokumentApi';
+import type { IBarnMedSamvær } from '../../../typer/dokumentApiBlankett';
 
 interface Props {
   barnMedSamvær: IBarnMedSamvær[];
