+++ src/server/blankett/components/NyttBarnSammePartner.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import AnnenForelder from './AnnenForelder';
-import type { IBarnMedSamvær } from '../../../typer/dokumentApi';
+import type { IBarnMedSamvær } from '../../../typer/dokumentApiBlankett';
 
 interface Props {
   barnMedSamvær: IBarnMedSamvær[];
