+++ src/server/components/Header.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { NavIkon } from './ikoner/navIkon';
-import type { Flettefelt } from '../../typer/dokumentApi';
+import type { Flettefelt } from '../../typer/dokumentApiBrev';
 import { validerFlettefelt } from '../utils/valideringer/validerFlettefelt';
 import { Maalform } from '../../typer/sanitygrensesnitt';
