+++ src/server/components/serializers/FlettefeltSerializer.tsx
@@ -1,4 +1,4 @@
-import type { Flettefelter } from '../../../typer/dokumentApi';
+import type { Flettefelter } from '../../../typer/dokumentApiBrev';
 import React from 'react';
 import { Feil } from '../../utils/Feil';
 import { validerFlettefelt } from '../../utils/valideringer/validerFlettefelt';
