+++ src/server/hentAvansertDokumentHtml.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import type { IBrevMedSignatur } from '../typer/dokumentApi';
+import type { IBrevMedSignatur } from '../typer/dokumentApiBrev';
 import AvansertDokument from './components/AvansertDokument';
 import type { Datasett } from './sanity/sanityClient';
 import { client } from './sanity/sanityClient';
