+++ src/server/søknadgenerator.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { renderToStaticMarkup } from 'react-dom/server';
-import type { ISøknad, IVerdiliste } from '../typer/dokumentApi';
+import type { ISøknad, IVerdiliste } from '../typer/dokumentApiBrev';
 import { dagensDatoFormatert } from './utils/util';
 import css from './utils/css';
 import søknadCSS from './utils/soknad-css';
